api_poc/gui/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
requests>=2.25.0
plotly>=5.0.0
numpy>=1.21.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.0
supabase>=1.0.0